requirements.txt
# ========================================
# Olist E-Commerce Data Pipeline
# Python Dependencies
# ========================================

# Cloud Storage
boto3>=1.34.0                   # AWS S3 client
botocore>=1.34.0                # AWS SDK core

# Database
psycopg2-binary>=2.9.9          # PostgreSQL adapter
sqlalchemy>=2.0.0               # SQL toolkit

# Data Processing
pandas>=2.1.0                   # Data manipulation
numpy>=1.26.0                   # Numerical operations
polars>=0.19.0                  # Fast DataFrame library
scipy>=1.11.0                   # Statistical analysis

# Data Source
kaggle>=1.6.0                   # Kaggle API client

# Configuration
python-dotenv>=1.0.0            # Environment variable management

# Orchestration
apache-airflow>=2.7.0           # Workflow orchestration (optional)

# Testing
pytest>=7.4.0                   # Testing framework
pytest-cov>=4.1.0               # Coverage reporting

# Code Quality
flake8>=6.1.0                   # Linting
black>=23.0.0                   # Code formatting

# Monitoring
prometheus-client>=0.18.0       # Metrics export (optional)
